Add AfterPortlet support to support the new version of the NewsBox extension in MW 1.39+
The positioning of this code snippet is chosen so that the NewsBox box is rendered in a similar style as all the other portlets, i.e. as it should be.
If this snippet were truly "after portlet", we'd end up with visual cruft that looks atrocious, so obviously I chose not to go down that route.

Change-Id: If2ad166c9fdafc04c74295d24ca337ddd0bd5174